fix: 사건 정보 요청 위치 변경
사건 정보 탭 -> 사건 상세 페이지
